feat: 가진 돈보다 비싼 물품 구매 시도 예외 처리
