+++ docs/README.md
@@ -37,7 +37,8 @@
   - [x] 구매할 상품이 없으면 예외를 반환한다.
   - [x] 구매할 상품이 남은 수량이 없으면 예외를 반환한다.
   - [x] 구매할 상품은 빈 값을 받을 수 없다.
-- [ ] 구매후 남은 금액을 보여준다.
+  - [x] 남은 돈보다 구매할 상품의 가격이 비싸면 구매할 수 없다.
+- [ ] 구매할 상품의 가격
 - [ ] 모든 제품의 수량이 0이 되거나, 수량이 남은 제품 중 가장 저렴한 제품 가격보다 남은 금액이 적으면 상품 구매를 종료한다.
 - [ ] 잔돈을 표시한다.
