+++ src/main/java/vendingmachine/validator/ProductValidator.java
@@ -1,12 +1,8 @@
 package vendingmachine.validator;
 
-import static vendingmachine.Constants.*;
-
 import java.util.List;
 
-import vendingmachine.model.ProductModel;
 import vendingmachine.service.ProductService;
-import vendingmachine.view.ErrorView;
 
 public class ProductValidator extends CommonValidator {
 	public static boolean checkProduct(String inputLine) {
@@ -18,35 +14,6 @@ public class ProductValidator extends CommonValidator {
 		} catch (IllegalArgumentException exception) {
 			return false;
 		}
-	}
-
-	public static boolean checkProductName(String inputLine) {
-		try {
-			exceptionStringEmptyOrSpace(inputLine);
-			exceptionNameNotMatch(inputLine);
-			exceptionHasZeroCount(inputLine);
-			return true;
-		} catch (IllegalArgumentException exception) {
-			ErrorView.error(ERROR_PRODUCTS);
-			return false;
-		}
-	}
-
-	private static void exceptionNameNotMatch(String inputLine) {
-		if (!ProductModel.hasProductName(inputLine)) {
-			throw new IllegalArgumentException();
-		}
-	}
-
-	private static void exceptionHasZeroCount(String inputLine) {
-		if (!ProductModel.hasMoreThanOneCount(inputLine)) {
-			throw new IllegalArgumentException();
-		}
-	}
-
-	private static void exceptionStringEmptyOrSpace(String inputLine) {
-		exceptionStringEmpty(inputLine);
-		exceptionStringSpace(inputLine);
 	}
 
 	private static void exceptionNotContainsBrackets(String inputLine) {
