+++ src/main/java/vendingmachine/view/ProductView.java
@@ -6,6 +6,7 @@ public class ProductView {
 	public static void messageInputProduct() {
 		System.out.println(MESSAGE_INPUT_PRODUCT);
 	}
+
 	public static void messageOrderProduct() {
 		System.out.println(MESSAGE_ORDER_PRODUCT);
 	}
